+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 # CHANGELOG
+
+## 2018.11.08-23:52
+
++ Added `BinarySearchTree`, `Node` and `BinarySearchTreeTest` java classes.
++ BST working on insertion.
++ BST working on pre, in and post order search.
++ Test only prints search example case.